Modification of several things.
Modified the Game loop so a desired FPS will be reached; this will make the cpu sleep as not to run too much while we don't want that.
Modified the AddComponent of the GameObject so that we can just use the this pointer; with that I also changed the constructor parameter of each component to a std::weak_ptr, as a shared_ptr is not really needed there.