Fixed Input and Commands
Made the Command an interface, added an Axis and Axis2D command, these have a proteced member var that stores the inputValue.
This inputValue can also be set.
Now we can inherit commands from this, and using an AI manager or something, we can set this input to control objects without keyboard/controller.
Added some filters to store base classes.
Removed a way to store the current input in the InputManager, als removed the way to ask those values. This will now be handled using the Axis and Axis2D commands.