Added physics
Added physics to the game.

When a GameObject has a RigidBody component, this will be affected by gravity. I made it so nothing will slide, this is a design feature (perhaps adding frictions later will fix this.)

MoveCommand has been updated to set the X/Y velocity of the RigidBody.
You can check whenever a RigidBody is grounded or not.

Grounded/Kinematic RigidBodies will not get their Y velocity updated.