Added basic sound support
Added a ServiceLocator where you can get a soundsystem from.
Added a soundsystem using SDL_mixer.
Added a decorator to this system.

Added some test sounds, 1 will play when you jump and another when the game starts.
Both are loaded in by default.

Loading and playing a sound happens on a different thread.