Modified SceneManager + Main.cpp + BoxCollisions
Modified the SceneManager:
The SceneManager can now go to the next or the previous Scene.
The SceneManager can now also remove/add scenes. (the AddScenes allows caching of a scene somewhere else while removing the scene and later re-adding the scene, this will change the order)
You can also now Set the active Scene.
And you can ask a pointer to the active Scene.

Added 2 commands to the game to be able to switch between scenes.

However since switching scenes can give certain problems, I added a check in the MoveCommand to compare the Scene Root with the Root Parent of the player.
Only if those match, the command will execute.
This required a function in GameObject and Scene to get this Root.

Cleaned up the Main.cpp:
Removed redundant stuff, so that this file is a bit more empty.
Modified the TextComponent to have a default " " string in it -> having nothing would crash the game.

Modified the BoxCollisions:
There is now finally layer support.
This means a BoxCollider can have a set layers to ignore collisions with.

In GridComponent added platform characters to the test string.
These have a different Collision layer.

When the player jumps, the ignore layers are set to ignore these platform layers. Resulting in the player being capable to jump up to different platforms.

This is still very WIP, yet a working proof of concept.
The biggest problem is that our player doesn't have any States yet, resulting in the un-ignoring of those platforms only happening when the player is holding down a movement key as this is handled in the MoveCommand.