Added LevelLoader
Added a LevelLoader Component.
You can pass a path to this component and when the CreateLevels() function is called, will go through all paths and create levels with the correct grid in them.

Noticed a major problem in the CollisionManager; collision is carried through levels, so added a check if the RootObject matches, only then we want to have some collisions.
This will take a big load on the game so can be a lot more optimized.