Added PlayerManager
Added PlayerManager and PlayerSpawnComponents.

The PlayerManager is an Observer to all Scenes.
When a scene is set to the active scene, they send notify observers with the SceneLoaded event.

When the PlayerManager receives this event, it will go through the level to find PlayerSpawnPoints, and spawn the player to the correct point.

This way the player is always in the grid.

Also modified GameObject.h, we can now have multiple components of the same type. We can also get all components with a type, or get all components in children with a type.

Modified GridComponent to be capable of spawning PlayerSpawnPoints.