Added Player Creation + Correct Removal
Added the Creation of players in the PlayerManager; if the PlayerCount isn't 0, it will spawn as many players as the count, as long as the count is less then or equal to the storedPlayerInfo.
When this Count is 0 and we have players, we simply mark them for delete and remove all links to them.

Modified GameObject and Scene(Manager) and the Update loop.
Now we go through all object in the scene every frame and if one is marked for delete, correctly remove them.
We do this recursivly so we check everything.