Added the Lives and Score displays
In the PlayerManager, added the Lives and Score display.
Renamed everything from Health to Lives, as we don't have Health but set Lives, we don't take damage, we straight up lose a life.

Added a way to remove Commands.
When players are destroyed, so are their Commands.
Added a file for GameEvents, not tied to the engine.

Modified the DamageCommand to handle Lives instead of Health, this command shouldn't be used so will be removed later.