Added PlayerComponent + Overlap Events
Added a PlayerComponent to each player.
This component will handle all game related stuff, events, states, you name it.
This is an observer and a subject.

Modified the BoxCollisions to store another set of layers; if anythings overlaps with the correct layers it will send out an event (ObjectOverlapping) with the GameObject that is overlapping as payload.

This can be used to check certain things; for now in the PlayerComponent , we check if we are overlapping with another player, if so we gain points but lose a life.

The PlayerComponent chains these events further to any other observers.

If the Player dies, we get the event, and we stop movement.  We now also have a bool for this in the BoxCollision; it will simply be ignored everytime.
This way we don't have to delete the player, as this can cause errors with the Commands.

Modified the PlayerManager; now we can respawn a player. This is thanks to the playerId that is now given at start.