Added Enemy
Added very basic Enemies.
From the ground up, these work the same as players; have their own spawnpoints, own manager,..
On scene load, any leftover enemies get destroyed, and we check the new level if there are any enemy spawnpoints, if there are, we fill those up.

Added support for the EnemySpawnpoints in the level files.

Modified the PhysicsManager + Rigidbody.
A RigidBody can now ignore gravity, this is basically the same as kinematic, but it goes through the update loop in RigidBody.

Also added the correct enemy behaviour; when it touches a player it "kills" them, sending the player back to their spawn point.