Added Shooting
Added a way to shoot; using a ShootCommand.

This Command checks if the player is running, and nothing else, then check if we can fire a bubble, if so, we do.

This is just a placeholder, as it doesn't spawn an object yet, but the functionallity is there including sound.

Removed 2 redundant Commands.