Added GameManager
Added a Working GameLoop.

The PlayerManager handles the events given from the Players.
When all players are dead or all enemies are dead, the GameManager get's notified to switch scenes.

It will switch to either the next scene or the last scene.
Eitherway, in the last scene, it will get the max score of the session, and will set this in the EndMenu.
That is the Score that will be able to be submitted to the highscore list.

Also added a check to the write so players can only write once per session.